refactor lightstep tracer tests into separate file
