+++ test/CMakeLists.txt
@@ -99,6 +99,7 @@ add_executable(envoy-test
   common/stats/statsd_test.cc
   common/stats/thread_local_store_test.cc
   common/tracing/http_tracer_impl_test.cc
+  common/tracing/lightstep_tracer_impl_test.cc
   common/upstream/cds_api_impl_test.cc
   common/upstream/cluster_manager_impl_test.cc
   common/upstream/health_checker_impl_test.cc
