+++ test/common/tracing/http_tracer_impl_test.cc
@@ -337,110 +337,6 @@ TEST(HttpTracerUtilityTest, operationTypeToString) {
   EXPECT_EQ("egress", HttpTracerUtility::toString(OperationName::Egress));
 }
 
-class LightStepDriverTest : public Test {
-public:
-  void setup(Json::Object& config, bool init_timer) {
-    std::unique_ptr<lightstep::TracerOptions> opts(new lightstep::TracerOptions());
-    opts->access_token = "sample_token";
-    opts->tracer_attributes["lightstep.component_name"] = "component";
-
-    ON_CALL(cm_, httpAsyncClientForCluster("fake_cluster"))
-        .WillByDefault(ReturnRef(cm_.async_client_));
-
-    if (init_timer) {
-      timer_ = new NiceMock<Event::MockTimer>(&tls_.dispatcher_);
-      EXPECT_CALL(*timer_, enableTimer(std::chrono::milliseconds(1000)));
-    }
-
-    driver_.reset(new LightStepDriver(config, cm_, stats_, tls_, runtime_, std::move(opts)));
-  }
-
-  void setupValidDriver() {
-    EXPECT_CALL(cm_, get("fake_cluster")).WillRepeatedly(Return(&cm_.thread_local_cluster_));
-    ON_CALL(*cm_.thread_local_cluster_.cluster_.info_, features())
-        .WillByDefault(Return(Upstream::ClusterInfo::Features::HTTP2));
-
-    std::string valid_config = R"EOF(
-      {"collector_cluster": "fake_cluster"}
-    )EOF";
-    Json::ObjectPtr loader = Json::Factory::LoadFromString(valid_config);
-
-    setup(*loader, true);
-  }
-
-  const std::string operation_name_{"test"};
-  Http::TestHeaderMapImpl request_headers_{
-      {":path", "/"}, {":method", "GET"}, {"x-request-id", "foo"}};
-  const Http::TestHeaderMapImpl response_headers_{{":status", "500"}};
-  SystemTime start_time_;
-  Http::AccessLog::MockRequestInfo request_info_;
-
-  std::unique_ptr<LightStepDriver> driver_;
-  NiceMock<Event::MockTimer>* timer_;
-  Stats::IsolatedStoreImpl stats_;
-  NiceMock<Upstream::MockClusterManager> cm_;
-  NiceMock<Runtime::MockRandomGenerator> random_;
-  NiceMock<Runtime::MockLoader> runtime_;
-  NiceMock<ThreadLocal::MockInstance> tls_;
-  NiceMock<LocalInfo::MockLocalInfo> local_info_;
-};
-
-TEST_F(LightStepDriverTest, InitializeDriver) {
-  {
-    std::string invalid_config = R"EOF(
-      {"fake" : "fake"}
-    )EOF";
-    Json::ObjectPtr loader = Json::Factory::LoadFromString(invalid_config);
-
-    EXPECT_THROW(setup(*loader, false), EnvoyException);
-  }
-
-  {
-    std::string empty_config = "{}";
-    Json::ObjectPtr loader = Json::Factory::LoadFromString(empty_config);
-
-    EXPECT_THROW(setup(*loader, false), EnvoyException);
-  }
-
-  {
-    // Valid config but not valid cluster.
-    EXPECT_CALL(cm_, get("fake_cluster")).WillOnce(Return(nullptr));
-
-    std::string valid_config = R"EOF(
-      {"collector_cluster": "fake_cluster"}
-    )EOF";
-    Json::ObjectPtr loader = Json::Factory::LoadFromString(valid_config);
-
-    EXPECT_THROW(setup(*loader, false), EnvoyException);
-  }
-
-  {
-    // Valid config, but upstream cluster does not support http2.
-    EXPECT_CALL(cm_, get("fake_cluster")).WillRepeatedly(Return(&cm_.thread_local_cluster_));
-    ON_CALL(*cm_.thread_local_cluster_.cluster_.info_, features()).WillByDefault(Return(0));
-
-    std::string valid_config = R"EOF(
-      {"collector_cluster": "fake_cluster"}
-    )EOF";
-    Json::ObjectPtr loader = Json::Factory::LoadFromString(valid_config);
-
-    EXPECT_THROW(setup(*loader, false), EnvoyException);
-  }
-
-  {
-    EXPECT_CALL(cm_, get("fake_cluster")).WillRepeatedly(Return(&cm_.thread_local_cluster_));
-    ON_CALL(*cm_.thread_local_cluster_.cluster_.info_, features())
-        .WillByDefault(Return(Upstream::ClusterInfo::Features::HTTP2));
-
-    std::string valid_config = R"EOF(
-      {"collector_cluster": "fake_cluster"}
-    )EOF";
-    Json::ObjectPtr loader = Json::Factory::LoadFromString(valid_config);
-
-    setup(*loader, true);
-  }
-}
-
 TEST(HttpNullTracerTest, BasicFunctionality) {
   HttpNullTracer null_tracer;
   MockConfig config;
@@ -491,158 +387,6 @@ TEST_F(HttpTracerImplTest, BasicFunctionalityNodeSet) {
   EXPECT_CALL(*span, setTag("node_id", "node_name"));
 
   tracer_->startSpan(config_, request_headers_, request_info_);
-}
-
-TEST_F(LightStepDriverTest, FlushSeveralSpans) {
-  setupValidDriver();
-
-  Http::MockAsyncClientRequest request(&cm_.async_client_);
-  Http::AsyncClient::Callbacks* callback;
-  const Optional<std::chrono::milliseconds> timeout(std::chrono::seconds(5));
-
-  EXPECT_CALL(cm_.async_client_, send_(_, _, timeout))
-      .WillOnce(
-          Invoke([&](Http::MessagePtr& message, Http::AsyncClient::Callbacks& callbacks,
-                     const Optional<std::chrono::milliseconds>&) -> Http::AsyncClient::Request* {
-            callback = &callbacks;
-
-            EXPECT_STREQ("/lightstep.collector.CollectorService/Report",
-                         message->headers().Path()->value().c_str());
-            EXPECT_STREQ("fake_cluster", message->headers().Host()->value().c_str());
-            EXPECT_STREQ("application/grpc", message->headers().ContentType()->value().c_str());
-
-            return &request;
-          }));
-
-  EXPECT_CALL(runtime_.snapshot_, getInteger("tracing.lightstep.min_flush_spans", 5))
-      .Times(2)
-      .WillRepeatedly(Return(2));
-  EXPECT_CALL(runtime_.snapshot_, getInteger("tracing.lightstep.request_timeout", 5000U))
-      .WillOnce(Return(5000U));
-
-  SpanPtr first_span = driver_->startSpan(request_headers_, operation_name_, start_time_);
-  first_span->finishSpan();
-
-  SpanPtr second_span = driver_->startSpan(request_headers_, operation_name_, start_time_);
-  second_span->finishSpan();
-
-  Http::MessagePtr msg(new Http::ResponseMessageImpl(
-      Http::HeaderMapPtr{new Http::TestHeaderMapImpl{{":status", "200"}}}));
-
-  msg->trailers(Http::HeaderMapPtr{new Http::TestHeaderMapImpl{{"grpc-status", "0"}}});
-
-  callback->onSuccess(std::move(msg));
-
-  EXPECT_EQ(1U, cm_.thread_local_cluster_.cluster_.info_->stats_store_
-                    .counter("grpc.lightstep.collector.CollectorService.Report.success")
-                    .value());
-
-  callback->onFailure(Http::AsyncClient::FailureReason::Reset);
-
-  EXPECT_EQ(1U, cm_.thread_local_cluster_.cluster_.info_->stats_store_
-                    .counter("grpc.lightstep.collector.CollectorService.Report.failure")
-                    .value());
-  EXPECT_EQ(2U, cm_.thread_local_cluster_.cluster_.info_->stats_store_
-                    .counter("grpc.lightstep.collector.CollectorService.Report.total")
-                    .value());
-  EXPECT_EQ(2U, stats_.counter("tracing.lightstep.spans_sent").value());
-}
-
-TEST_F(LightStepDriverTest, FlushSpansTimer) {
-  setupValidDriver();
-
-  const Optional<std::chrono::milliseconds> timeout(std::chrono::seconds(5));
-  EXPECT_CALL(cm_.async_client_, send_(_, _, timeout));
-
-  EXPECT_CALL(runtime_.snapshot_, getInteger("tracing.lightstep.min_flush_spans", 5))
-      .WillOnce(Return(5));
-
-  SpanPtr span = driver_->startSpan(request_headers_, operation_name_, start_time_);
-  span->finishSpan();
-
-  // Timer should be re-enabled.
-  EXPECT_CALL(*timer_, enableTimer(std::chrono::milliseconds(1000)));
-  EXPECT_CALL(runtime_.snapshot_, getInteger("tracing.lightstep.request_timeout", 5000U))
-      .WillOnce(Return(5000U));
-  EXPECT_CALL(runtime_.snapshot_, getInteger("tracing.lightstep.flush_interval_ms", 1000U))
-      .WillOnce(Return(1000U));
-
-  timer_->callback_();
-
-  EXPECT_EQ(1U, stats_.counter("tracing.lightstep.timer_flushed").value());
-  EXPECT_EQ(1U, stats_.counter("tracing.lightstep.spans_sent").value());
-}
-
-TEST_F(LightStepDriverTest, FlushOneSpanGrpcFailure) {
-  setupValidDriver();
-
-  Http::MockAsyncClientRequest request(&cm_.async_client_);
-  Http::AsyncClient::Callbacks* callback;
-  const Optional<std::chrono::milliseconds> timeout(std::chrono::seconds(5));
-
-  EXPECT_CALL(cm_.async_client_, send_(_, _, timeout))
-      .WillOnce(
-          Invoke([&](Http::MessagePtr& message, Http::AsyncClient::Callbacks& callbacks,
-                     const Optional<std::chrono::milliseconds>&) -> Http::AsyncClient::Request* {
-            callback = &callbacks;
-
-            EXPECT_STREQ("/lightstep.collector.CollectorService/Report",
-                         message->headers().Path()->value().c_str());
-            EXPECT_STREQ("fake_cluster", message->headers().Host()->value().c_str());
-            EXPECT_STREQ("application/grpc", message->headers().ContentType()->value().c_str());
-
-            return &request;
-          }));
-  EXPECT_CALL(runtime_.snapshot_, getInteger("tracing.lightstep.min_flush_spans", 5))
-      .WillOnce(Return(1));
-  EXPECT_CALL(runtime_.snapshot_, getInteger("tracing.lightstep.request_timeout", 5000U))
-      .WillOnce(Return(5000U));
-
-  SpanPtr span = driver_->startSpan(request_headers_, operation_name_, start_time_);
-  span->finishSpan();
-
-  Http::MessagePtr msg(new Http::ResponseMessageImpl(
-      Http::HeaderMapPtr{new Http::TestHeaderMapImpl{{":status", "200"}}}));
-
-  // No trailers, gRPC is considered failed.
-  callback->onSuccess(std::move(msg));
-
-  EXPECT_EQ(1U, cm_.thread_local_cluster_.cluster_.info_->stats_store_
-                    .counter("grpc.lightstep.collector.CollectorService.Report.failure")
-                    .value());
-  EXPECT_EQ(1U, cm_.thread_local_cluster_.cluster_.info_->stats_store_
-                    .counter("grpc.lightstep.collector.CollectorService.Report.total")
-                    .value());
-  EXPECT_EQ(1U, stats_.counter("tracing.lightstep.spans_sent").value());
-}
-
-TEST_F(LightStepDriverTest, SerializeAndDeserializeContext) {
-  setupValidDriver();
-
-  // Supply bogus context, that will be simply ignored.
-  const std::string invalid_context = "notvalidcontext";
-  request_headers_.insertOtSpanContext().value(invalid_context);
-  driver_->startSpan(request_headers_, operation_name_, start_time_);
-
-  std::string injected_ctx = request_headers_.OtSpanContext()->value().c_str();
-  EXPECT_FALSE(injected_ctx.empty());
-
-  // Supply empty context.
-  request_headers_.removeOtSpanContext();
-  SpanPtr span = driver_->startSpan(request_headers_, operation_name_, start_time_);
-
-  injected_ctx = request_headers_.OtSpanContext()->value().c_str();
-  EXPECT_FALSE(injected_ctx.empty());
-
-  // Context can be parsed fine.
-  lightstep::BinaryCarrier ctx;
-  std::string context = Base64::decode(injected_ctx);
-  ctx.ParseFromString(context);
-
-  // Supply parent context, request_headers has properly populated x-ot-span-context.
-  SpanPtr span_with_parent = driver_->startSpan(request_headers_, operation_name_, start_time_);
-  injected_ctx = request_headers_.OtSpanContext()->value().c_str();
-  EXPECT_FALSE(injected_ctx.empty());
 }
 
 } // Tracing
